Add OpenMeteo API composables
This adds a composable which fetches and processes data from OpenMeteo's
servers. It uses the singleton pattern to ensure only one set of data is
retrieved. It is substantially rewritten from the example provided in
the OpenMeteo docs to allow for strict typescript rules and to work
within a vue composable.