Refactor chart component into multiple composables
This is a major refactoring of the `SymptomCharts` component.
previously, all logic and templating was contained within the component.
Now, each chart has its own composable for logic, while templating is
handled by the generic bar chart component and the SymptomCharts
component. Chart data and options are passed down to the bar chart
components by props.

This refactoring means that as more charts are added, the SymptomCharts
component does not become too large and take on too much responsibility.