Add CSS utility classes for range inputs
Range inputs are used in the application for handling some number inputs.
These CSS improvements make the use of the sliders easier to understand
and more prominent in the UI.

I decided to use a colour gradient to display sliders, to indicate
intensity: from yellow to red. The sliders are also larger in height and
show a pointer when used on the desktop. Slider width has been increased
to the full width of the containing component.