Separate text location from input in GeoLocation
The text input and current text location (returned by Nominatim) were
tightly coupled with a `v-model`. This caused issues where updating the
text input also updated the outputs in graphs etc.

This commit separates the two. The text input is passed as a query to
the geocoding API. When the confirmedLocationName is found, the text
input is updated.

It also renames a number of variables to reflect their new roles.