Synchronise text updates with state handler
Due to th async nature of reverseGeocoding, the button text was updated
at slightly different times. This caused the UI to feel unresponsive.

This commit fixes this issue with a state handler `locationSource`:
the location is either handled by GPS, MANUAL, or NONE (initial). This
is used in `watch(confirmedLocationName)` to synchronise the text updates.