Rewrite `usePollenFilters` as modular functions
`usePollenFilters` has been totally rewritten, now it has a function
`applyFilters` which accepts `PollenData` and a list of filter functions.
It applies the filter functions one by one to each `PollenRecord` in
`PollenData`.

The filter functions are higher-order function which take a filtering
argument and return a function which processes the `PollenRecord`

This commit adds `typeFilter`, `levelFilter` and `timeframeFilter`,
whcih filter by pollen type, pollen level, and timeframe, respectively.